Move Lookup class into import_mixins.
