+++ openslides_backend/action/mixins/import_mixins.py
@@ -381,3 +381,33 @@ class JsonUploadMixin(BaseImportJsonUpload):
                             f"Unknown type in conversion: type:{type_} is_object:{str(is_object)} is_list:{str(is_list)}"
                         )
         super().validate_instance(instance)
+
+
+class Lookup:
+    def __init__(
+        self,
+        datastore: DatastoreService,
+        collection: str,
+        names: List[str],
+        field: str = "name",
+    ) -> None:
+        if not names:
+            self.name_to_id: Dict[str, Optional[int]] = {}
+        else:
+            self.name_to_id = {
+                entry[field]: entry["id"]
+                for entry in datastore.filter(
+                    collection,
+                    Or(*[FilterOperator(field, "=", name) for name in names]),
+                    ["id", field],
+                ).values()
+            }
+
+    def check_duplicate(self, name: str) -> bool:
+        result = name in self.name_to_id
+        if not result:
+            self.name_to_id[name] = None
+        return result
+
+    def get_id_by_name(self, name: str) -> Optional[int]:
+        return self.name_to_id.get(name)
